Update and rename Resource Readers/Writers.
Changing the current converters by prefixing them with 'Binary[...]'. Planning to have other converters for formats such as SJSON. This may be a messy, unnecessary way of doing things but I've been thinking for a bit of a better way and I can't, so this is how it's going to be, at least for now.