Fix Assimp bone extraction.
You ever do something one way and then try and brute force it that way because you can't understand why it won't work... and then come back after hours and days and realize you're just dumb?

Yeah.

Anyway, using the IndexBuffer to index into the VertexBuffers for the weights (like, y'know, every other type of VertexBuffer) so we assign the bones *properly*. Still not 100% on this however, as some bones seem to be missing. Could be due to assigning the current mesh the bones? I'm not entirely certain, or I'd fix it lul.